refactor: rename category types
