+++ src/cmd/pdf/post-pdf-template-builder.ts
@@ -2,8 +2,8 @@ import { Post } from '@/model/post'
 import { PostFileMapManager } from '@/service/post/post-file-map'
 import fs from 'fs'
 import { BlogSettingService } from '@/service/blog-setting'
-import { PostCatService } from '@/service/post/post-category'
-import { PostCat } from '@/model/post-category'
+import { PostCatService } from '@/service/post/post-cat'
+import { PostCat } from '@/model/post-cat'
 import { markdownItFactory } from '@cnblogs/markdown-it-presets'
 import { AuthManager } from '@/auth/auth-manager'
