+++ src/cmd/post-cat/del-selected-cat.ts
@@ -1,7 +1,7 @@
 import { ProgressLocation, window } from 'vscode'
-import { PostCatService } from '@/service/post/post-category'
+import { PostCatService } from '@/service/post/post-cat'
 import { Alert } from '@/infra/alert'
-import { PostCat } from '@/model/post-category'
+import { PostCat } from '@/model/post-cat'
 import { extTreeViews } from '@/tree-view/tree-view-register'
 import { PostCatTreeItem } from '@/tree-view/model/post-category-tree-item'
 import { postCategoryDataProvider } from '@/tree-view/provider/post-category-tree-data-provider'
