+++ src/cmd/post-cat/update-post-cat-treeview.ts
@@ -1,9 +1,9 @@
 import fs from 'fs'
 import { ProgressLocation, window, Uri, workspace } from 'vscode'
-import { PostCat } from '@/model/post-category'
-import { PostCatService } from '@/service/post/post-category'
+import { PostCat } from '@/model/post-cat'
+import { PostCatService } from '@/service/post/post-cat'
 import { inputPostCat } from './input-post-cat'
-import { PostCatCfg } from '@/ctx/cfg/post-category'
+import { PostCatCfg } from '@/ctx/cfg/post-cat'
 import { WorkspaceCfg } from '@/ctx/cfg/workspace'
 import { Alert } from '@/infra/alert'
 import { PostCatTreeItem } from '@/tree-view/model/post-category-tree-item'
