+++ src/cmd/show-local-file-to-post-info.ts
@@ -2,7 +2,7 @@ import path from 'path'
 import { Uri } from 'vscode'
 import { Alert } from '@/infra/alert'
 import { PostService } from '@/service/post/post'
-import { PostCatService } from '@/service/post/post-category'
+import { PostCatService } from '@/service/post/post-cat'
 import { PostFileMapManager } from '@/service/post/post-file-map'
 import { searchPostByTitle } from '@/service/post/search-post-by-title'
 import { viewPostOnline } from './view-post-online'
