+++ src/service/post/post-cfg-panel.ts
@@ -2,7 +2,7 @@ import { cloneDeep } from 'lodash-es'
 import vscode, { WebviewPanel, Uri } from 'vscode'
 import { Post } from '@/model/post'
 import { globalCtx } from '@/ctx/global-ctx'
-import { PostCatService } from './post-category'
+import { PostCatService } from './post-cat'
 import { PostTagService } from './post-tag'
 import { PostService } from './post'
 import { WebviewMsg } from '@/model/webview-msg'
