+++ src/tree-view/convert.ts
@@ -2,7 +2,7 @@ import format from 'date-fns/format'
 import { homedir } from 'os'
 import { MarkdownString, ThemeIcon, TreeItem, TreeItemCollapsibleState, Uri } from 'vscode'
 import { Post } from '@/model/post'
-import { PostCat } from '@/model/post-category'
+import { PostCat } from '@/model/post-cat'
 import { PostFileMapManager } from '@/service/post/post-file-map'
 import { BaseTreeItemSource } from '@/tree-view/model/base-tree-item-source'
 import { IconThemeCfg } from '@/ctx/cfg/icon-theme'
