+++ src/tree-view/model/category-list-tree-item.ts
@@ -1,9 +1,9 @@
-import { PostCat } from '@/model/post-category'
+import { PostCat } from '@/model/post-cat'
 import { PostCatTreeItem } from './post-category-tree-item'
 import { PostEntryMetadata, PostMetadata, PostTagMetadata } from './post-metadata'
 import { PostTreeItem } from './post-tree-item'
 
-export type PostCategoriesListTreeItem =
+export type PostCatListTreeItem =
     | PostCat
     | PostCatTreeItem
     | PostTreeItem
